lambda/requirements.txt
playwright==1.44.0
Pillow>=10.0.0
pytesseract>=0.3.10
